requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Airflow
apache-airflow==2.7.3

# Database (if querying Airflow DB)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# HTTP requests (for Airflow REST API)
requests==2.31.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0

# Logging & monitoring
python-logging-loki==0.3.2